"check_style_score -> 3 of 4 passed. style score: 0/1 "

+++ ADS1-assignment/m24/1/Solution.java
@@ -37,7 +37,7 @@ public final class Solution {
                 // String ans = Arrays.toString(hash.get(token[1]));
                 // String[] answer = ans.split(",");
                 try {
-                    if (Integer.parseInt(token[2]) == 1) {
+                    if (Double.parseDouble(token[2]) == 1) {
                         System.out.println(hash.get(token[1])[0]);
                     } else if (Double.parseDouble(token[2]) == 2) {
                         System.out.println(hash.get(token[1])[1]);
